# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for متتبع الإزالة (Remover Tracker)
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type LogEntryOutcome =
  (typeof LogEntryOutcome)[keyof typeof LogEntryOutcome];

export const LogEntryOutcome = {
  Held: "Held",
  Partial: "Partial",
  Caved: "Caved",
} as const;

export interface LogEntry {
  id: number;
  date: string;
  what: string;
  source: string;
  category: string;
  /** @nullable */
  hoursRecovered?: number | null;
  outcome: LogEntryOutcome;
  /** @nullable */
  note?: string | null;
  createdAt: string;
}

export type CreateLogEntryBodyOutcome =
  (typeof CreateLogEntryBodyOutcome)[keyof typeof CreateLogEntryBodyOutcome];

export const CreateLogEntryBodyOutcome = {
  Held: "Held",
  Partial: "Partial",
  Caved: "Caved",
} as const;

export interface CreateLogEntryBody {
  date: string;
  what: string;
  source: string;
  category: string;
  /** @nullable */
  hoursRecovered?: number | null;
  outcome: CreateLogEntryBodyOutcome;
  /** @nullable */
  note?: string | null;
}

export type UpdateLogEntryBodyOutcome =
  (typeof UpdateLogEntryBodyOutcome)[keyof typeof UpdateLogEntryBodyOutcome];

export const UpdateLogEntryBodyOutcome = {
  Held: "Held",
  Partial: "Partial",
  Caved: "Caved",
} as const;

export interface UpdateLogEntryBody {
  date?: string;
  what?: string;
  source?: string;
  category?: string;
  /** @nullable */
  hoursRecovered?: number | null;
  outcome?: UpdateLogEntryBodyOutcome;
  /** @nullable */
  note?: string | null;
}

export interface DailyItem {
  id: number;
  date: string;
  what: string;
  /** @nullable */
  replacement?: string | null;
  riskLevel: number;
  done: boolean;
  createdAt: string;
}

export interface CreateDailyItemBody {
  date: string;
  what: string;
  /** @nullable */
  replacement?: string | null;
  riskLevel: number;
}

export interface UpdateDailyItemBody {
  done?: boolean;
  what?: string;
  /** @nullable */
  replacement?: string | null;
  riskLevel?: number;
}

export interface MasterRule {
  id: number;
  category: string;
  what: string;
  /** @nullable */
  why?: string | null;
  createdAt: string;
}

export interface CreateMasterRuleBody {
  category: string;
  what: string;
  /** @nullable */
  why?: string | null;
}

export interface Priorities {
  date: string;
  /** @nullable */
  priority1?: string | null;
  /** @nullable */
  priority2?: string | null;
  /** @nullable */
  priority3?: string | null;
}

export interface IfThenPlan {
  id: number;
  date: string;
  ifCondition: string;
  thenAction: string;
  createdAt: string;
}

export interface CreateIfThenPlanBody {
  date: string;
  ifCondition: string;
  thenAction: string;
}

export interface DashboardStats {
  totalNos: number;
  weekNos: number;
  hoursRecovered: number;
  commitmentRate: number;
  streak: number;
}

export interface DayActivity {
  date: string;
  count: number;
}

export interface CategoryCount {
  category: string;
  count: number;
}