# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for متتبع الإزالة (Remover Tracker)
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all log entries
 */
export const GetLogEntriesResponseItem = zod.object({
  id: zod.number(),
  date: zod.string(),
  what: zod.string(),
  source: zod.string(),
  category: zod.string(),
  hoursRecovered: zod.number().nullish(),
  outcome: zod.enum(["Held", "Partial", "Caved"]),
  note: zod.string().nullish(),
  createdAt: zod.string(),
});
export const GetLogEntriesResponse = zod.array(GetLogEntriesResponseItem);

/**
 * @summary Create a new log entry
 */
export const CreateLogEntryBody = zod.object({
  date: zod.string(),
  what: zod.string(),
  source: zod.string(),
  category: zod.string(),
  hoursRecovered: zod.number().nullish(),
  outcome: zod.enum(["Held", "Partial", "Caved"]),
  note: zod.string().nullish(),
});

/**
 * @summary Update a log entry
 */
export const UpdateLogEntryParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateLogEntryBody = zod.object({
  date: zod.string().optional(),
  what: zod.string().optional(),
  source: zod.string().optional(),
  category: zod.string().optional(),
  hoursRecovered: zod.number().nullish(),
  outcome: zod.enum(["Held", "Partial", "Caved"]).optional(),
  note: zod.string().nullish(),
});

export const UpdateLogEntryResponse = zod.object({
  id: zod.number(),
  date: zod.string(),
  what: zod.string(),
  source: zod.string(),
  category: zod.string(),
  hoursRecovered: zod.number().nullish(),
  outcome: zod.enum(["Held", "Partial", "Caved"]),
  note: zod.string().nullish(),
  createdAt: zod.string(),
});

/**
 * @summary Delete a log entry
 */
export const DeleteLogEntryParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Get daily items for today
 */
export const GetDailyItemsResponseItem = zod.object({
  id: zod.number(),
  date: zod.string(),
  what: zod.string(),
  replacement: zod.string().nullish(),
  riskLevel: zod.number(),
  done: zod.boolean(),
  createdAt: zod.string(),
});
export const GetDailyItemsResponse = zod.array(GetDailyItemsResponseItem);

/**
 * @summary Create a daily item
 */
export const CreateDailyItemBody = zod.object({
  date: zod.string(),
  what: zod.string(),
  replacement: zod.string().nullish(),
  riskLevel: zod.number(),
});

/**
 * @summary Update a daily item (e.g. toggle done)
 */
export const UpdateDailyItemParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateDailyItemBody = zod.object({
  done: zod.boolean().optional(),
  what: zod.string().optional(),
  replacement: zod.string().nullish(),
  riskLevel: zod.number().optional(),
});

export const UpdateDailyItemResponse = zod.object({
  id: zod.number(),
  date: zod.string(),
  what: zod.string(),
  replacement: zod.string().nullish(),
  riskLevel: zod.number(),
  done: zod.boolean(),
  createdAt: zod.string(),
});

/**
 * @summary Delete a daily item
 */
export const DeleteDailyItemParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Get permanent ban rules
 */
export const GetMasterRulesResponseItem = zod.object({
  id: zod.number(),
  category: zod.string(),
  what: zod.string(),
  why: zod.string().nullish(),
  createdAt: zod.string(),
});
export const GetMasterRulesResponse = zod.array(GetMasterRulesResponseItem);

/**
 * @summary Create a master rule
 */
export const CreateMasterRuleBody = zod.object({
  category: zod.string(),
  what: zod.string(),
  why: zod.string().nullish(),
});

/**
 * @summary Delete a master rule
 */
export const DeleteMasterRuleParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Get today's priorities
 */
export const GetPrioritiesResponse = zod.object({
  date: zod.string(),
  priority1: zod.string().nullish(),
  priority2: zod.string().nullish(),
  priority3: zod.string().nullish(),
});

/**
 * @summary Save today's priorities
 */
export const SavePrioritiesBody = zod.object({
  date: zod.string(),
  priority1: zod.string().nullish(),
  priority2: zod.string().nullish(),
  priority3: zod.string().nullish(),
});

export const SavePrioritiesResponse = zod.object({
  date: zod.string(),
  priority1: zod.string().nullish(),
  priority2: zod.string().nullish(),
  priority3: zod.string().nullish(),
});

/**
 * @summary Get all if/then plans
 */
export const GetIfThenPlansResponseItem = zod.object({
  id: zod.number(),
  date: zod.string(),
  ifCondition: zod.string(),
  thenAction: zod.string(),
  createdAt: zod.string(),
});
export const GetIfThenPlansResponse = zod.array(GetIfThenPlansResponseItem);

/**
 * @summary Create an if/then plan
 */
export const CreateIfThenPlanBody = zod.object({
  date: zod.string(),
  ifCondition: zod.string(),
  thenAction: zod.string(),
});

/**
 * @summary Delete an if/then plan
 */
export const DeleteIfThenPlanParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Get dashboard KPI stats
 */
export const GetDashboardStatsResponse = zod.object({
  totalNos: zod.number(),
  weekNos: zod.number(),
  hoursRecovered: zod.number(),
  commitmentRate: zod.number(),
  streak: zod.number(),
});

/**
 * @summary Get last 14 days activity (daily counts)
 */
export const GetDashboardActivityResponseItem = zod.object({
  date: zod.string(),
  count: zod.number(),
});
export const GetDashboardActivityResponse = zod.array(
  GetDashboardActivityResponseItem,
);

/**
 * @summary Get log entries grouped by category
 */
export const GetDashboardByCategoryResponseItem = zod.object({
  category: zod.string(),
  count: zod.number(),
});
export const GetDashboardByCategoryResponse = zod.array(
  GetDashboardByCategoryResponseItem,
);